fix: if/else scope in assert statements
A statement like `assert x == 1 if False else "foobar"` will
be evaluated to assert "foobar", which is always True. This
means that checking for a response status code like
`assert status_code == 200 if i < 5 else 429` will be evaluated
to `assert status_code == 200` for i < 5, but also to
`assert 429` for i >= 5 and not `assert status_code == 429`.

The fix is to get the right scope for the if/else condition by using
parentheses, so that the right response status code will be checked
if i >= 5.